requirements.txt
aiogram==3.24.0
aiohttp==3.13.3
aiosqlite==0.22.1
alembic==1.18.3
asyncpg==0.31.0
loguru==0.7.3
python-dotenv==1.2.1
redis==5.0.1
SQLAlchemy==2.0.46
typing_extensions>=4.9.0
aio_pika==9.4.0
pandas==2.3.2
openpyxl==3.1.5
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pytest>=8.0.0
pytest-asyncio>=0.24.0
